feat: major enhance
Handle on back pressed, if the yes button is clicked a leave message is sent to the server and to the adapter. Refactored notifyAdapterChange to comport the scroll function, messages are now updated through a mutable live data, no longer needing the client viewmodel to have an instance of the recycler view nor the adapter. Audio recorder name is now globally saved so the view can play the same audio file when clicked, because the audio name is the epoch, having two calls to the getEpoch function will have minor differences in name, and the user will not be able to play the correct audio. checkDisconnected has been refactored to a single function, if the server is not down it'll handle the same way it did, this saved 3 repetitive and is no longer too error prone.